# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.HttpAggregateFunction:
    enum:
    - Minimum
    - Maximum
    - Average
    - Sum
    - Count
    type: string
    x-enum-varnames:
    - Minimum
    - Maximum
    - Average
    - Sum
    - Count
  main.HttpError:
    properties:
      inner_message:
        type: string
      message:
        type: string
    type: object
  main.HttpPartialResult:
    properties:
      aggregation:
        type: string
      count:
        description: Count associated with the result.
        type: integer
      double_value:
        description: Double value (nullable).
        type: number
      float_value:
        description: Float value (nullable).
        type: number
      is_null:
        description: Indicates if the result is null.
        type: boolean
      result_type:
        description: 'Type of result: "INT", "FLOAT", "DOUBLE".'
        type: string
      value:
        description: Integer value (nullable).
        type: integer
    type: object
  main.HttpQueryRequest:
    properties:
      group_columns:
        items:
          type: string
        type: array
      select:
        items:
          $ref: '#/definitions/main.HttpSelect'
        type: array
      table_name:
        type: string
    type: object
  main.HttpQueryResponse:
    properties:
      error:
        $ref: '#/definitions/main.HttpError'
      values:
        items:
          $ref: '#/definitions/main.HttpValue'
        type: array
    type: object
  main.HttpResult:
    properties:
      processing_time:
        type: integer
      result:
        $ref: '#/definitions/main.HttpQueryResponse'
    type: object
  main.HttpSelect:
    properties:
      column:
        type: string
      function:
        $ref: '#/definitions/main.HttpAggregateFunction'
    type: object
  main.HttpValue:
    properties:
      grouping_value:
        type: string
      results:
        items:
          $ref: '#/definitions/main.HttpPartialResult'
        type: array
    type: object
  main.ParquetColumnInfo:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact: {}
  title: Swagger Distributed data aggregation system API
  version: "1.0"
paths:
  /query:
    post:
      consumes:
      - application/json
      description: Queries data with specified grouping and selection
      parameters:
      - description: Query Request
        in: body
        name: query
        required: true
        schema:
          $ref: '#/definitions/main.HttpQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Query has been processed
          schema:
            $ref: '#/definitions/main.HttpResult'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Query data from a table
      tags:
      - query
  /status:
    get:
      description: Checks controller status
      produces:
      - application/json
      responses:
        "200":
          description: Health check passed
          schema:
            type: string
      summary: Health check endpoint
      tags:
      - health check
  /tables:
    get:
      description: Returns the names of non-empty folders in the data path
      produces:
      - application/json
      responses:
        "200":
          description: List of table names
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List available tables
      tags:
      - tables
  /tables/columns:
    get:
      description: Returns column names and their types for a given table. Filters
        out columns of unsupported types.
      parameters:
      - description: Table name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of columns with their types
          schema:
            items:
              items:
                $ref: '#/definitions/main.ParquetColumnInfo'
              type: array
            type: array
        "400":
          description: Bad request with error message
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get table columns
      tags:
      - tables
  /tables/select-columns:
    get:
      description: Returns column names and their types for a given table. Filters
        out columns of types unsupported for aggregations.
      parameters:
      - description: Table name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of columns with their types
          schema:
            items:
              items:
                $ref: '#/definitions/main.ParquetColumnInfo'
              type: array
            type: array
        "400":
          description: Bad request with error message
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get table columns that can be aggregated
      tags:
      - tables
swagger: "2.0"